archive/pr-resolution-archive/requirements-graphql.txt
# PR Resolution Automation System Requirements
# Additional dependencies beyond the base setup

# GraphQL and API
graphene>=3.3.0
graphene[sqlalchemy]>=3.3.0
graphql-core>=3.2.0

# Database
neo4j>=5.15.0
redis>=5.0.0
asyncpg>=0.29.0

# Caching and Performance
aiocache>=0.12.0
asyncio-throttle>=1.0.0
psutil>=5.9.0

# API Testing
httpx>=0.27.0
pytest-asyncio>=0.21.0

# Monitoring and Observability
prometheus-client>=0.19.0
structlog>=23.1.0

# OpenAI Integration
openai>=1.3.0
tiktoken>=0.5.0

# AI/ML Libraries
langchain>=0.1.0
pydantic-ai>=0.1.0

# Background Task Processing
celery>=5.3.0
dramatiq>=1.15.0

# AI Testing and Validation
pytest-httpx>=0.27.0
pytest-mock>=3.12.0
factory-boy>=3.3.0

# Additional utilities
pydantic-settings>=2.0.0
python-multipart>=0.0.6
asyncio-throttle>=1.0.0